Update coordinator.py to improve the "Last Update" sensor
Update coordinator.py to try to improve  the code for the "Last Update" sensor. Trying to keep the last date and time the device is Online and also trying to not to update the date and time when the API blocks the connection due to limit API Calls and did't send any data.