number entity to manage timedelta and sensor
Create a number entity with time slider to dynamically change the timedelta, set by default at 30 minutes.  The number will be located at the "Control" sections inside the integration. And also create a sensor that shows the state of number entity. Sensor that will be located at the "Sensors" section inside the integration, so it could be used to show  the state of  the number entity in the frontend and avoid involuntary changes of the timedelta.